+++ prbt_hardware_support/test/unittests/unittest_operation_mode_setup_executor.cpp
@@ -22,6 +22,8 @@
 
 #include <ros/time.h>
 
+#include <pilz_msgs/GetSpeedOverride.h>
+
 #include <prbt_hardware_support/OperationModes.h>
 #include <prbt_hardware_support/operation_mode_setup_executor.h>
 #include <prbt_hardware_support/operation_mode_setup_executor_node_service_calls.h>
@@ -286,8 +288,8 @@ TEST_P(OperationModeSetupExecutorTestSpeedOverrideNice, testSpeedOverride)
 
   executor_->updateOperationMode(op_mode);
 
-  auto req = GetSpeedOverrideRequest();
-  auto res = GetSpeedOverrideResponse();
+  auto req = pilz_msgs::GetSpeedOverrideRequest();
+  auto res = pilz_msgs::GetSpeedOverrideResponse();
   executor_->getSpeedOverride(req, res);
   EXPECT_EQ(res.speed_override, GetParam().second);
 }